+++ R/print.easycorrelation.R
@@ -21,31 +21,33 @@ print.easycorrelation <- function(x, digits = 2, stars = TRUE, ...) {
   formatted_table$n_Obs <- NULL
 
   cat(insight::export_table(formatted_table,
-                            format = "text",
-                            caption = table_caption,
-                            footer = .print_easycorrelation_add_footer(x)))
+    format = "text",
+    caption = table_caption,
+    footer = .print_easycorrelation_add_footer(x)
+  ))
   invisible(x)
 }
 
 
 
 #' @keywords internal
-.print_easycorrelation_add_footer <- function(x){
-
+.print_easycorrelation_add_footer <- function(x) {
   footer <- ""
 
   # P-adjust
   if (isFALSE(attributes(x)$bayesian)) {
-    footer <- paste0(footer,
-                     "\np-value adjustment method: ",
-                     parameters::format_p_adjust(attributes(x)$p_adjust))
+    footer <- paste0(
+      footer,
+      "\np-value adjustment method: ",
+      parameters::format_p_adjust(attributes(x)$p_adjust)
+    )
   }
 
   # N-obs
   if (!is.null(x$n_Obs)) {
     if (length(unique(x$n_Obs)) == 1) {
       nobs <- unique(x$n_Obs)
-    } else{
+    } else {
       nobs <- paste0(min(x$n_Obs), "-", max(x$n_Obs))
     }
     footer <- paste0(footer, "\nObservations: ", nobs, "\n")